:memos: レビューを受けての修正

+++ ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: shwatana <shwatana@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/17 19:45:16 by shwatana          #+#    #+#             */
-/*   Updated: 2022/04/06 15:46:15 by shwatana         ###   ########.fr       */
+/*   Updated: 2022/04/12 11:02:10 by shwatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 	i = 0;
 	while (s[i] != '\0')
 	{
-		f(i, s + i);
+		f((unsigned int)i, s + i);
 		i++;
 	}
 }
